Clarify jni netcdf4 public API.
