+++ cdm/radial/src/test/java/ucar/nc2/TestNc4Structures.java
@@ -16,7 +16,7 @@ import org.junit.rules.TemporaryFolder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ucar.ma2.InvalidRangeException;
-import ucar.nc2.jni.netcdf.Nc4Iosp;
+import ucar.nc2.jni.NetcdfClibrary;
 import ucar.nc2.util.CancelTask;
 import ucar.nc2.write.NetcdfCopier;
 import ucar.nc2.write.NetcdfFileFormat;
@@ -42,7 +42,7 @@ public class TestNc4Structures {
     // Ignore this class's tests if NetCDF-4 isn't present.
     // We're using @Before because it shows these tests as being ignored.
     // @BeforeClass shows them as *non-existent*, which is not what we want.
-    Assume.assumeTrue("NetCDF-4 C library not present.", Nc4Iosp.isClibraryPresent());
+    Assume.assumeTrue("NetCDF-4 C library not present.", NetcdfClibrary.isClibraryPresent());
   }
 
   @Test
