+++ netcdf4/src/main/java/ucar/nc2/jni/netcdf/package.html
@@ -1,0 +1,15 @@
+<!--
+  ~ Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
+  ~  See LICENSE for license information.
+  -->
+
+<HTML>
+<HEAD>
+  <TITLE>package ucar.nc2.jni.netcdf</TITLE>
+</HEAD>
+<BODY>
+
+DO NOT USE, internal to the Java-NetCDF library.
+
+</BODY>
+</HTML>
