+++ netcdf4/src/main/java/ucar/nc2/jni/package.html
@@ -1,0 +1,14 @@
+<!--
+  ~ Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
+  ~  See LICENSE for license information.
+  -->
+<HTML>
+<HEAD>
+  <TITLE>package ucar.nc2.jni</TITLE>
+</HEAD>
+<BODY>
+
+Public interface to the NetCDF C library.
+
+</BODY>
+</HTML>
